fix: blogs alignment and add blogs data
